+++ Movie Repository/Entities/Rating.cs
@@ -5,7 +5,7 @@ namespace Movie_Repository.Entities
     public class Rating
     {
         public Guid Id { get; set; }
-        public ushort Value { get; set; }
+        public int Value { get; set; }
 
         public Guid MovieId { get; set; }
         public virtual Movie Movie { get; set; }
