+++ Repository API/DTO/RatingDTO.cs
@@ -5,6 +5,6 @@ namespace Repository_API.DTO
     public class RatingDTO
     {
         public UserDTO User{ get; set; }
-        public ushort Value { get; set; }
+        public int Value { get; set; }
     }
 }
